requirementnts.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
django-bootstrap4==24.4
django-crispy-forms==2.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
jsonfield==3.1.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.6.0
soupsieve==2.6
sqlparse==0.5.2
swapper==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
